Clean up some config tests
